+++ packages/backtest/src/analyzers/Returns.ts
@@ -1,22 +1,6 @@
 import dayjs from 'dayjs'
 import BaseAnalyzer from './BaseAnalyzer'
 
-function formatNumber(val) {
-  if (val >= 1000) {
-    const formattedValue = (val / 1000).toFixed(1);
-    return `$${formattedValue}k`;
-  } else {
-
-    const op = {
-      style: 'currency',
-      currency: 'USD',
-      maximumFractionDigits: 3
-    }
-
-    return val.toLocaleString('en-US', op)
-  }
-}
-
 class RetursAnalyzer extends BaseAnalyzer {
   name = 'returns'
   data = [] as any[]