+++ packages/cli/src/api/index.ts
@@ -2,10 +2,12 @@ import cache from './cache'
 import code from './code'
 import data from './data'
 import auth from './login'
+import report from './report'
 
 export {
   cache,
   code,
   data,
-  auth
+  auth,
+  report
 }