+++ packages/backtest/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @zapcli/backtest
 
+## 1.0.3
+
+### Patch Changes
+
+- add reports command
+- Updated dependencies
+  - @zapcli/core@1.0.3
+
 ## 1.0.2
 
 ### Patch Changes