+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @zapcli/cli
 
+## 1.0.3
+
+### Patch Changes
+
+- add reports command
+- Updated dependencies
+  - @zapcli/backtest@1.0.3
+  - @zapcli/core@1.0.3
+
 ## 1.0.2
 
 ### Patch Changes