+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zapcli/core
 
+## 1.0.3
+
+### Patch Changes
+
+- add reports command
+
 ## 1.0.2
 
 ### Patch Changes